Avoid calling ssm_create() from ship_apply_tag() if ssm_index < 0.
This would happen if a TAG level 3 weapon had no specified SSM (or an invalid one). Since nothing can call ssm_create() with a negative ssm_index now, changed it to a size_t and simplified one of the verification conditionals.

+++ code/hud/hudartillery.h
@@ -78,7 +78,7 @@ void hud_artillery_update();
 void hud_artillery_render();
 
 // start a subspace missile effect
-void ssm_create(object *target, vec3d *start, int ssm_index, ssm_firing_info *override, int team);
+void ssm_create(object *target, vec3d *start, size_t ssm_index, ssm_firing_info *override, int team);
 
 // Goober5000
 extern int ssm_info_lookup(const char *name);
